products rev4

+++ assets/js/productsList.js
@@ -1,3 +1,20 @@
+// window.addEventListener("load", ()=>{
+//   const loader = document.querySelector(".loader")
+//   setTimeout(()=>{
+// opac
+
+
+
+
+//   })
+//   loader.classList.add("loader-hidden")
+//   loader.addEventListener("transitioned", ()=>{
+//     document.body.removeChild("loader")
+//   })
+// })
+
+
+
 import {
   getAllData,
   getDataById,
@@ -25,9 +42,8 @@ function printTable(products) {
                 <td>${product.category}</td>
                 <td>${product.title}</td>
                 <td>${product.price}</td>
-                <td>${product.rating.count}</td>
                 <td>
-                <button type="button" class="btn btn-outline-success"style="border: none; font-weight: bold;" >Delete</button>
+                <a href="detail.html?id=${product.id}" ><button type="button" class="btn btn-outline-success about"style="border: none; font-weight: bold;">About</button></a>
                 </td>   
                 <td>
                 <button data-id="${product.id}" type="button" class="btn btn-outline-danger delete"style="border: none; font-weight: bold;" >Delete</button>
@@ -40,13 +56,29 @@ function printTable(products) {
   const allDeleteBtn = document.querySelectorAll(".delete");
   allDeleteBtn.forEach((btn) => {
     btn.addEventListener("click", (e) => {
-     const deletedId =e.target.getAttribute("data-id");
-     
-     btn.closest("tr").remove()
-      deleteDataById(endpoint.products,deletedId)
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+      }).then((result) => {
+        if (result.isConfirmed) {
+          console.log(result)
+          const deletedId =e.target.getAttribute("data-id");
+          deleteDataById(endpoint.products,deletedId)
+          e.target.closest("tr").remove()
+          Swal.fire({
+            title: "Deleted!",
+            text: "Your file has been deleted.",
+            icon: "success"
+          });
+        }
+      });
+    
     });
   });
 }
 
-
-deleteDataById(endpoint.products,2)
